requirements.txt
groq
qdrant-client
sentence-transformers
streamlit
langchain-text-splitters
pymupdf
tqdm
python-dotenv
